refactor: update TypeScript configuration and improve theme components
- Updated TypeScript paths for better readability.
- Changed the reference link in next-env.d.ts for TypeScript documentation.
- Replaced SideBar with ProfileSidebar in the layout for improved user profile display.
- Simplified Header component by removing unnecessary elements and added space for future content.
- Introduced a new ProfileSidebar component for displaying user profile information.
